Add files via upload
PingMonitor as its described is exactly that, Designed in a spreadsheet style with as yet completed Email notification
function for out of hours email notification if an address is unreachable after a predetermined number of ping is missed.
It includes an adjustable  timer for Ping intervals. 
Save command to store all IP or  Domain Names and Email Address, Ping Timer period , 
to enable reloading of Addresses and Set Timer interval etc. if program crashes or day to day program use. 
Stats Button to search logged data for individual IP/Names giving successful and Failed ping data as well as Mean return times.